// src/app/models/srm/common/payment-negotiation.ts
export class PaymentNegotiation {
    paymentConditionId: number=-1;
    discount: number=0;
    paymentDeadlinesDays: number=0;
    negotiationTypeId: number=-1;
    currencyTypeId: number=-1;
    paymentCurrencyId: number=-1;
    paymentRateTypeId: number=-1;
    exchangeRateId:number=-1;
    exchangeRateValue: number=0;
    paymentExchangeRate: number=0;
    paymentMethod: number=-1;
    taxPlanId: number=-1;
    idDiscountType:number=-1;
    name:string="";
    observations:string=""
}